backend/pyproject.toml
[project]
name = "fuzzy-poetry-api"
version = "0.1.0"
description = "The FastAPI backend for Fuzzy Poetry"
authors = [
    {name = "rskarp",email = "rkarp1997@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13, <4.0"
dependencies = [
    "fastapi (>=0.121.1,<0.122.0)",
    "uvicorn (>=0.38.0,<0.39.0)",
    "openai (>=2.7.2,<3.0.0)",
    "lancedb (>=0.25.3,<0.26.0)",
    "nltk (>=3.9.2,<4.0.0)",
    "pillow (>=12.0.0,<13.0.0)",
    "python-datamuse (>=2.0.1,<3.0.0)",
    "regex (>=2025.11.3,<2026.0.0)",
    "boto3 (>=1.40.72,<2.0.0)",
    "open-clip-torch>=3.2.0",
    "python-dotenv>=1.2.1",
    "mangum>=0.20.0",
]


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
